@@ -100,12 +100,12 @@
     "prismjs": "^1.21.0",
     "prop-types": "^15.6.0",
     "query-string": "6.6.0",
-    "react": "16.12.0",
+    "react": "16.14.0",
     "react-autosize-textarea": "7.1.0",
     "react-bootstrap": "^0.32.0",
     "react-date-range": "^1.0.0-beta",
     "react-document-title": "2.0.3",
-    "react-dom": "16.12.0",
+    "react-dom": "16.14.0",
     "react-keydown": "^1.9.7",
     "react-lazyload": "^2.3.0",
     "react-mentions": "4.0.1",
