@@ -1,40 +1,70 @@
-import $ from 'jquery';
 import React from 'react';
-import PropTypes from 'prop-types';
+import {RouteComponentProps} from 'react-router/lib/Router';
 import DocumentTitle from 'react-document-title';
 
+import {Client} from 'app/api';
 import withApi from 'app/utils/withApi';
 import LazyLoad from 'app/components/lazyLoad';
 import withOrganization from 'app/utils/withOrganization';
-import SentryTypes from 'app/sentryTypes';
+import {Organization, Project} from 'app/types';
 
-class OrganizationStatsContainer extends React.Component {
-  static propTypes = {
-    api: PropTypes.object.isRequired,
-    routes: PropTypes.array.isRequired,
-    organization: SentryTypes.Organization.isRequired,
+import {ProjectTotal, Point, OrgTotal} from './types';
+
+type Props = {
+  api: Client;
+  organization: Organization;
+} & RouteComponentProps<{orgId: string}, {}>;
+
+type RawData = {
+  received: Point[];
+  rejected: Point[];
+  blacklisted: Point[];
+};
+
+type State = {
+  projectsError: boolean;
+  projectsLoading: boolean;
+  projectsRequestsPending: number;
+  statsError: boolean;
+  statsLoading: boolean;
+  statsRequestsPending: number;
+  projectMap: null | Record<string, Project>;
+  rawProjectData: {
+    received: Record<string, Point[]>;
+    rejected: Record<string, Point[]>;
+    blacklisted: Record<string, Point[]>;
   };
+  rawOrgData: RawData;
+  orgStats: null | Point[];
+  orgTotal: null | OrgTotal;
+  projectTotals: null | ProjectTotal[];
+  querySince: number;
+  queryUntil: number;
+  pageLinks: null | string;
+};
 
-  constructor(props) {
+class OrganizationStatsContainer extends React.Component<Props, State> {
+  constructor(props: Props) {
     super(props);
     const until = Math.floor(new Date().getTime() / 1000);
     const since = until - 3600 * 24 * 7;
 
     this.state = {
-      projectsError: null,
+      projectsError: false,
       projectsLoading: false,
       projectsRequestsPending: 0,
-      statsError: null,
+      statsError: false,
       statsLoading: false,
       statsRequestsPending: 0,
       projectMap: null,
-      rawProjectData: {received: null, rejected: null, blacklisted: null},
-      rawOrgData: {received: null, rejected: null, blacklisted: null},
+      rawProjectData: {received: {}, rejected: {}, blacklisted: {}},
+      rawOrgData: {received: [], rejected: [], blacklisted: []},
       orgStats: null,
       orgTotal: null,
       projectTotals: null,
       querySince: since,
       queryUntil: until,
+      pageLinks: null,
     };
   }
 
@@ -42,7 +72,7 @@ class OrganizationStatsContainer extends React.Component {
     this.fetchData();
   }
 
-  UNSAFE_componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps: Props) {
     // If query string changes, it will be due to pagination.
     // Intentionally only fetch projects since stats are fetched for a fixed period during
     // the initial payload
@@ -55,7 +85,7 @@ class OrganizationStatsContainer extends React.Component {
     }
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps: Props) {
     const prevParams = prevProps.params,
       currentParams = this.props.params;
 
@@ -85,13 +115,13 @@ class OrganizationStatsContainer extends React.Component {
     this.props.api.request(this.getOrganizationProjectsEndpoint(), {
       query: this.props.location.query,
       success: (data, _textStatus, jqxhr) => {
-        const projectMap = {};
-        data.forEach(project => {
+        const projectMap: Record<string, Project> = {};
+        data.forEach((project: Project) => {
           projectMap[project.id] = project;
         });
 
         this.setState(prevState => ({
-          pageLinks: jqxhr.getResponseHeader('Link'),
+          pageLinks: jqxhr ? jqxhr.getResponseHeader('Link') : null,
           projectMap,
           projectsRequestsPending: prevState.projectsRequestsPending - 1,
         }));
@@ -116,7 +146,7 @@ class OrganizationStatsContainer extends React.Component {
 
     const statEndpoint = this.getOrganizationStatsEndpoint();
 
-    $.each(this.state.rawOrgData, statName => {
+    Object.keys(this.state.rawOrgData).forEach(statName => {
       this.props.api.request(statEndpoint, {
         query: {
           since: this.state.querySince,
@@ -143,7 +173,7 @@ class OrganizationStatsContainer extends React.Component {
       });
     });
 
-    $.each(this.state.rawProjectData, statName => {
+    Object.keys(this.state.rawProjectData).forEach(statName => {
       this.props.api.request(statEndpoint, {
         query: {
           since: this.state.querySince,
@@ -187,10 +217,11 @@ class OrganizationStatsContainer extends React.Component {
     let oReceived = 0;
     let oRejected = 0;
     let oBlacklisted = 0;
-    const orgPoints = []; // accepted, rejected, blacklisted
+    const orgPoints: Point[] = []; // accepted, rejected, blacklisted
     const aReceived = [0, 0]; // received, points
     const rawOrgData = this.state.rawOrgData;
-    $.each(rawOrgData.received, (idx, point) => {
+
+    rawOrgData.received.forEach((point, idx) => {
       const dReceived = point[1];
       const dRejected = rawOrgData.rejected[idx][1];
       const dBlacklisted = rawOrgData.blacklisted[idx][1];
@@ -210,11 +241,12 @@ class OrganizationStatsContainer extends React.Component {
     this.setState({
       orgStats: orgPoints,
       orgTotal: {
+        id: '',
         received: oReceived,
         rejected: oRejected,
         blacklisted: oBlacklisted,
         accepted: Math.max(0, oReceived - oRejected - oBlacklisted),
-        avgRate: aReceived[1] ? parseInt(aReceived[0] / aReceived[1] / 60, 10) : 0,
+        avgRate: aReceived[1] ? Math.round(aReceived[0] / aReceived[1] / 60) : 0,
       },
       statsLoading: false,
     });
@@ -222,12 +254,13 @@ class OrganizationStatsContainer extends React.Component {
 
   processProjectData() {
     const rawProjectData = this.state.rawProjectData;
-    const projectTotals = [];
-    $.each(rawProjectData.received, (projectId, data) => {
+    const projectTotals: ProjectTotal[] = [];
+    Object.keys(rawProjectData.received).forEach(projectId => {
+      const data = rawProjectData.received[projectId];
       let pReceived = 0;
       let pRejected = 0;
       let pBlacklisted = 0;
-      $.each(data, (idx, point) => {
+      data.forEach((point, idx) => {
         pReceived += point[1];
         pRejected += rawProjectData.rejected[projectId][idx][1];
         pBlacklisted += rawProjectData.blacklisted[projectId][idx][1];
