@@ -55,9 +55,7 @@ export async function fetchOrgMembers(
   return [];
 }
 
-type IndexedMembersByProject = {
-  [key: string]: Member['user'][];
-};
+type IndexedMembersByProject = Record<string, Member['user'][]>;
 
 /**
  * Convert a list of members with user & project data
