@@ -11,12 +11,13 @@ import SentryTypes from 'app/sentryTypes';
 import theme from 'app/utils/theme';
 import withApi from 'app/utils/withApi';
 import withOrganization from 'app/utils/withOrganization';
+import parseLinkHeader from 'app/utils/parseLinkHeader';
 import {escape} from 'app/utils';
 import {formatVersion} from 'app/utils/formatters';
 
 // This is not an exported action/function because releases list uses AsyncComponent
 // and this is not re-used anywhere else afaict
-function getOrganizationReleases(api, organization, conditions = null) {
+async function getOrganizationReleases(api, organization, conditions = null) {
   const query = {};
   Object.keys(conditions).forEach(key => {
     let value = conditions[key];
@@ -28,7 +29,8 @@ function getOrganizationReleases(api, organization, conditions = null) {
     }
   });
   api.clear();
-  return api.requestPromise(`/organizations/${organization.slug}/releases/`, {
+  return api.requestPromise(`/organizations/${organization.slug}/releases/stats/`, {
+    includeAllArgs: true,
     method: 'GET',
     query,
   });
@@ -86,7 +88,7 @@ class ReleaseSeries extends React.Component {
     this.props.api.clear();
   }
 
-  fetchData() {
+  async fetchData() {
     const {api, organization, projects, environments, period, start, end} = this.props;
     const conditions = {
       start,
@@ -95,15 +97,33 @@ class ReleaseSeries extends React.Component {
       environment: environments,
       statsPeriod: period,
     };
-    getOrganizationReleases(api, organization, conditions)
-      .then(releases => {
+    let hasMore = true;
+    const releases = [];
+    while (hasMore) {
+      try {
+        const [newReleases, , xhr] = await getOrganizationReleases(
+          api,
+          organization,
+          conditions
+        );
+        releases.push(...newReleases);
         if (this._isMounted) {
           this.setReleasesWithSeries(releases);
         }
-      })
-      .catch(() => {
+
+        const pageLinks = xhr && xhr.getResponseHeader('Link');
+        if (pageLinks) {
+          const paginationObject = parseLinkHeader(pageLinks);
+          hasMore = paginationObject && paginationObject.next.results;
+          conditions.cursor = paginationObject.next.cursor;
+        } else {
+          hasMore = false;
+        }
+      } catch {
         addErrorMessage(t('Error fetching releases'));
-      });
+        hasMore = false;
+      }
+    }
   }
 
   setReleasesWithSeries(releases) {
@@ -155,7 +175,7 @@ class ReleaseSeries extends React.Component {
           show: false,
         },
         data: releases.map(release => ({
-          xAxis: +new Date(release.dateCreated),
+          xAxis: +new Date(release.date),
           name: formatVersion(release.version, true),
           value: formatVersion(release.version, true),
           onClick: () => {
