@@ -26,6 +26,7 @@ import {
   isGapSpan,
   isOrphanSpan,
   isEventFromBrowserJavaScriptSDK,
+  toPercent,
 } from './utils';
 import {DragManagerChildrenProps} from './dragManager';
 import SpanGroup from './spanGroup';
@@ -33,6 +34,8 @@ import {SpanRowMessage} from './styles';
 import * as DividerHandlerManager from './dividerHandlerManager';
 import {FilterSpans} from './traceView';
 import {ActiveOperationFilter} from './filter';
+import MeasurementsPanel from './measurementsPanel';
+import * as MeasurementsManager from './measurementsManager';
 
 type RenderedSpanTree = {
   spanTree: JSX.Element | null;
@@ -331,17 +334,23 @@ class SpanTree extends React.Component<PropType> {
     };
   };
 
-  renderRootSpan = (): RenderedSpanTree => {
+  generateBounds() {
     const {dragProps, trace} = this.props;
 
-    const rootSpan: RawSpanType = generateRootSpan(trace);
-
-    const generateBounds = boundsGenerator({
+    return boundsGenerator({
       traceStartTimestamp: trace.traceStartTimestamp,
       traceEndTimestamp: trace.traceEndTimestamp,
       viewStart: dragProps.viewWindowStart,
       viewEnd: dragProps.viewWindowEnd,
     });
+  }
+
+  renderRootSpan = (): RenderedSpanTree => {
+    const {trace} = this.props;
+
+    const rootSpan: RawSpanType = generateRootSpan(trace);
+
+    const generateBounds = this.generateBounds();
 
     return this.renderSpan({
       isRoot: true,
@@ -357,6 +366,48 @@ class SpanTree extends React.Component<PropType> {
       previousSiblingEndTimestamp: undefined,
     });
   };
+
+  renderSecondaryPanel() {
+    const {organization, event} = this.props;
+
+    if (!organization.features.includes('measurements')) {
+      return null;
+    }
+
+    const hasMeasurements = Object.keys(event.measurements ?? {}).length > 0;
+
+    // only display the secondary header if there are any measurements
+    if (!hasMeasurements) {
+      return null;
+    }
+
+    return (
+      <DividerHandlerManager.Consumer>
+        {(
+          dividerHandlerChildrenProps: DividerHandlerManager.DividerHandlerManagerChildrenProps
+        ) => {
+          const {dividerPosition} = dividerHandlerChildrenProps;
+
+          return (
+            <SecondaryHeader>
+              <div
+                style={{
+                  // the width of this component is shrunk to compensate for half of the width of the divider line
+                  width: `calc(${toPercent(dividerPosition)} - 0.5px)`,
+                }}
+              />
+              <DividerSpacer />
+              <MeasurementsPanel
+                event={event}
+                generateBounds={this.generateBounds()}
+                dividerPosition={dividerPosition}
+              />
+            </SecondaryHeader>
+          );
+        }}
+      </DividerHandlerManager.Consumer>
+    );
+  }
 
   render() {
     const {
@@ -374,10 +425,13 @@ class SpanTree extends React.Component<PropType> {
 
     return (
       <DividerHandlerManager.Provider interactiveLayerRef={this.traceViewRef}>
-        <TraceViewContainer ref={this.traceViewRef}>
-          {spanTree}
-          {infoMessage}
-        </TraceViewContainer>
+        <MeasurementsManager.Provider>
+          {this.renderSecondaryPanel()}
+          <TraceViewContainer ref={this.traceViewRef}>
+            {spanTree}
+            {infoMessage}
+          </TraceViewContainer>
+        </MeasurementsManager.Provider>
       </DividerHandlerManager.Provider>
     );
   }
@@ -387,6 +441,17 @@ const TraceViewContainer = styled('div')`
   overflow-x: hidden;
   border-bottom-left-radius: 3px;
   border-bottom-right-radius: 3px;
+`;
+
+const SecondaryHeader = styled('div')`
+  background-color: ${p => p.theme.gray100};
+  display: flex;
+
+  border-bottom: 1px solid ${p => p.theme.gray400};
+`;
+
+const DividerSpacer = styled('div')`
+  width: 1px;
 `;
 
 export default SpanTree;
