@@ -9,15 +9,11 @@ import {Panel} from 'app/components/panels';
 import space from 'app/styles/space';
 import Tooltip from 'app/components/tooltip';
 import {IconFire} from 'app/icons';
-import {WEB_VITAL_DETAILS} from 'app/views/performance/transactionVitals/constants';
+import {
+  WEB_VITAL_DETAILS,
+  LONG_WEB_VITAL_NAMES,
+} from 'app/views/performance/transactionVitals/constants';
 import {formattedValue} from 'app/utils/measurements/index';
-
-// translate known short form names into their long forms
-const LONG_MEASUREMENT_NAMES = Object.fromEntries(
-  Object.values(WEB_VITAL_DETAILS).map(value => {
-    return [value.slug, value.name];
-  })
-);
 
 type Props = {
   organization: Organization;
@@ -58,14 +54,14 @@ class RealUserMonitoring extends React.Component<Props> {
       const currentValue = formattedValue(record, value);
       const thresholdValue = formattedValue(record, record?.failureThreshold ?? 0);
 
-      if (!LONG_MEASUREMENT_NAMES.hasOwnProperty(name)) {
+      if (!LONG_WEB_VITAL_NAMES.hasOwnProperty(name)) {
         return null;
       }
 
       return (
         <div key={name}>
           <StyledPanel failedThreshold={failedThreshold}>
-            <Name>{LONG_MEASUREMENT_NAMES[name] ?? name}</Name>
+            <Name>{LONG_WEB_VITAL_NAMES[name] ?? name}</Name>
             <ValueRow>
               {failedThreshold ? (
                 <WarningIconContainer size="sm">
