@@ -1,14 +1,13 @@
 import {browserHistory} from 'react-router';
+import {RouteComponentProps} from 'react-router/lib/Router';
 import Cookies from 'js-cookie';
-import PropTypes from 'prop-types';
 import React from 'react';
-import Reflux from 'reflux';
 import classNames from 'classnames';
-import createReactClass from 'create-react-class';
 import isEqual from 'lodash/isEqual';
 import pickBy from 'lodash/pickBy';
 import * as qs from 'query-string';
 import {withProfiler} from '@sentry/react';
+import {Location} from 'history';
 
 import {Client} from 'app/api';
 import {DEFAULT_QUERY, DEFAULT_STATS_PERIOD} from 'app/constants';
@@ -30,15 +29,23 @@ import LoadingError from 'app/components/loadingError';
 import LoadingIndicator from 'app/components/loadingIndicator';
 import Pagination from 'app/components/pagination';
 import ProcessingIssueList from 'app/components/stream/processingIssueList';
-import SentryTypes from 'app/sentryTypes';
+import {
+  GlobalSelection,
+  Member,
+  Organization,
+  SavedSearch,
+  TagCollection,
+} from 'app/types';
 import StreamGroup from 'app/components/stream/group';
 import StreamManager from 'app/utils/streamManager';
 import parseApiError from 'app/utils/parseApiError';
 import parseLinkHeader from 'app/utils/parseLinkHeader';
+import withApi from 'app/utils/withApi';
 import withGlobalSelection from 'app/utils/withGlobalSelection';
 import withOrganization from 'app/utils/withOrganization';
 import withSavedSearches from 'app/utils/withSavedSearches';
 import withIssueTags from 'app/utils/withIssueTags';
+import {callIfFunction} from 'app/utils/callIfFunction';
 
 import IssueListActions from './actions';
 import IssueListFilters from './filters';
@@ -53,49 +60,77 @@ const DEFAULT_GRAPH_STATS_PERIOD = '24h';
 const STATS_PERIODS = new Set(['14d', '24h']);
 const DYNAMIC_COUNTS_STATS_PERIODS = new Set(['14d', '24h', 'auto']);
 
-const IssueListOverview = createReactClass({
-  displayName: 'IssueListOverview',
+type Params = {
+  orgId: string;
+};
 
-  propTypes: {
-    organization: SentryTypes.Organization,
-    selection: SentryTypes.GlobalSelection,
-    savedSearch: SentryTypes.SavedSearch,
-    savedSearches: PropTypes.arrayOf(SentryTypes.SavedSearch),
-    savedSearchLoading: PropTypes.bool.isRequired,
-    tags: PropTypes.object,
-  },
+type Props = {
+  api: Client;
+  location: Location;
+  organization: Organization;
+  params: Params;
+  selection: GlobalSelection;
+  savedSearch: SavedSearch;
+  savedSearches: SavedSearch[];
+  savedSearchLoading: boolean;
+  tags: TagCollection;
+} & RouteComponentProps<{searchId?: string}, {}>;
 
-  mixins: [Reflux.listenTo(GroupStore, 'onGroupChange')],
+type State = {
+  groupIds: string[];
+  selectAllActive: boolean;
+  realtimeActive: boolean;
+  pageLinks: string;
+  queryCount: number;
+  queryMaxCount: number;
+  error: string | null;
+  isSidebarVisible: boolean;
+  issuesLoading: boolean;
+  tagsLoading: boolean;
+  memberList: ReturnType<typeof indexMembersByProject>;
+  query?: string;
+};
 
-  getInitialState() {
+type EndpointParams = Partial<GlobalSelection['datetime']> & {
+  project: number[];
+  environment: string[];
+  query?: string;
+  sort?: string;
+  statsPeriod?: string;
+  groupStatsPeriod?: string;
+  cursor?: string;
+  page?: number | string;
+};
+
+class IssueListOverview extends React.Component<Props, State> {
+  constructor(props) {
+    super(props);
+
     const realtimeActiveCookie = Cookies.get('realtimeActive');
     const realtimeActive =
       typeof realtimeActiveCookie === 'undefined'
         ? false
         : realtimeActiveCookie === 'true';
 
-    return {
+    this.state = {
       groupIds: [],
       selectAllActive: false,
       realtimeActive,
       pageLinks: '',
-      queryCount: null,
-      error: false,
+      queryCount: 0,
+      queryMaxCount: 0,
+      error: null,
       isSidebarVisible: false,
       issuesLoading: true,
       tagsLoading: true,
       memberList: {},
-      // the project for the selected issues
-      // Will only be set if selected issues all belong
-      // to one project.
-      selectedProject: null,
     };
-  },
+  }
 
   componentDidMount() {
-    this.api = new Client();
-    this._streamManager = new StreamManager(GroupStore);
+    const links = parseLinkHeader(this.state.pageLinks);
     this._poller = new CursorPoller({
+      endpoint: links.previous?.href || '',
       success: this.onRealtimePoll,
     });
 
@@ -105,9 +140,9 @@ const IssueListOverview = createReactClass({
     // Start by getting searches first so if the user is on a saved search
     // or they have a pinned search we load the correct data the first time.
     this.fetchSavedSearches();
-  },
+  }
 
-  componentDidUpdate(prevProps, prevState) {
+  componentDidUpdate(prevProps: Props, prevState: State) {
     // Fire off profiling/metrics first
     if (prevState.issuesLoading && !this.state.issuesLoading) {
       // First Meaningful Paint for /organizations/:orgId/issues/
@@ -174,7 +209,7 @@ const IssueListOverview = createReactClass({
     ) {
       this.fetchData();
     } else if (
-      !this.lastRequest &&
+      !this._lastRequest &&
       prevState.issuesLoading === false &&
       this.state.issuesLoading
     ) {
@@ -182,14 +217,13 @@ const IssueListOverview = createReactClass({
       // This can happen when transitionTo is called
       this.fetchData();
     }
-  },
+  }
 
   componentWillUnmount() {
     this._poller.disable();
-    this.projectCache = {};
     GroupStore.reset();
-    this.api.clear();
-
+    this.props.api.clear();
+    callIfFunction(this.listener);
     // Reset store when unmounting because we always fetch on mount
     // This means if you navigate away from stream and then back to stream,
     // this component will go from:
@@ -201,45 +235,42 @@ const IssueListOverview = createReactClass({
     // cause weird side effects (e.g. ProcessingIssueList mounting and making
     // a request, but immediately unmounting when fetching saved searches)
     resetSavedSearches();
-  },
+  }
 
-  // Memoize projects fetched as selections are made
-  // This data is fed into the action toolbar for release data.
-  projectCache: {},
+  private _poller: any;
+  private _lastRequest: any;
+  private _streamManager = new StreamManager(GroupStore);
 
-  getQuery() {
+  getQuery(): string {
     if (this.props.savedSearch) {
       return this.props.savedSearch.query;
     }
 
     const {query} = this.props.location.query;
-    return typeof query === 'undefined' ? DEFAULT_QUERY : query;
-  },
+    return typeof query === 'undefined' ? DEFAULT_QUERY : (query as string);
+  }
 
-  getSort() {
-    return this.props.location.query.sort || DEFAULT_SORT;
-  },
+  getSort(): string {
+    return (this.props.location.query.sort as string) || DEFAULT_SORT;
+  }
 
-  getDefaultGroupStatsPeriod() {
-    return this.props.organization.features.includes('dynamic-issue-counts')
-      ? 'auto'
-      : DEFAULT_GRAPH_STATS_PERIOD;
-  },
-
-  getGroupStatsPeriod() {
-    const currentPeriod = this.props.location.query.groupStatsPeriod;
+  getGroupStatsPeriod(): string {
+    const currentPeriod =
+      typeof this.props.location.query?.groupStatsPeriod === 'string'
+        ? this.props.location.query?.groupStatsPeriod
+        : DEFAULT_GRAPH_STATS_PERIOD;
     return (this.props.organization.features.includes('dynamic-issue-counts')
       ? DYNAMIC_COUNTS_STATS_PERIODS
       : STATS_PERIODS
     ).has(currentPeriod)
       ? currentPeriod
       : DEFAULT_GRAPH_STATS_PERIOD;
-  },
+  }
 
-  getEndpointParams() {
+  getEndpointParams = (): EndpointParams => {
     const {selection} = this.props;
 
-    const params = {
+    const params: EndpointParams = {
       project: selection.projects,
       environment: selection.environments,
       query: this.getQuery(),
@@ -268,43 +299,41 @@ const IssueListOverview = createReactClass({
     }
 
     // only include defined values.
-    return pickBy(params, v => defined(v));
-  },
+    return pickBy(params, v => defined(v)) as EndpointParams;
+  };
 
-  getFeatures() {
-    return new Set(this.props.organization.features);
-  },
-
-  getGlobalSearchProjectIds() {
+  getGlobalSearchProjectIds = () => {
     return this.props.selection.projects;
-  },
+  };
 
   fetchMemberList() {
     const projectIds = this.getGlobalSearchProjectIds();
 
-    fetchOrgMembers(this.api, this.props.organization.slug, projectIds).then(members => {
-      this.setState({memberList: indexMembersByProject(members)});
-    });
-  },
+    fetchOrgMembers(this.props.api, this.props.organization.slug, projectIds).then(
+      members => {
+        this.setState({memberList: indexMembersByProject(members)});
+      }
+    );
+  }
 
   fetchTags() {
     const {organization, selection} = this.props;
     this.setState({tagsLoading: true});
-    loadOrganizationTags(this.api, organization.slug, selection).then(() =>
+    loadOrganizationTags(this.props.api, organization.slug, selection).then(() =>
       this.setState({tagsLoading: false})
     );
-  },
+  }
 
-  fetchData() {
+  fetchData = () => {
     GroupStore.loadInitialData([]);
 
     this.setState({
       issuesLoading: true,
-      queryCount: null,
-      error: false,
+      queryCount: 0,
+      error: null,
     });
 
-    const requestParams = {
+    const requestParams: any = {
       ...this.getEndpointParams(),
       limit: MAX_ITEMS,
       shortIdLookup: 1,
@@ -320,20 +349,24 @@ const IssueListOverview = createReactClass({
       requestParams.statsPeriod = DEFAULT_STATS_PERIOD;
     }
 
-    if (this.lastRequest) {
-      this.lastRequest.cancel();
+    if (this._lastRequest) {
+      this._lastRequest.cancel();
     }
 
     this._poller.disable();
 
-    this.lastRequest = this.api.request(this.getGroupListEndpoint(), {
+    this._lastRequest = this.props.api.request(this.getGroupListEndpoint(), {
       method: 'GET',
       data: qs.stringify(requestParams),
       success: (data, _, jqXHR) => {
+        if (!jqXHR) {
+          return;
+        }
+
         const {orgId} = this.props.params;
         // If this is a direct hit, we redirect to the intended result directly.
         if (jqXHR.getResponseHeader('X-Sentry-Direct-Hit') === '1') {
-          let redirect;
+          let redirect: string;
           if (data[0] && data[0].matchingEventId) {
             const {id, matchingEventId} = data[0];
             redirect = `/organizations/${orgId}/issues/${id}/events/${matchingEventId}/`;
@@ -356,13 +389,17 @@ const IssueListOverview = createReactClass({
         const pageLinks = jqXHR.getResponseHeader('Link');
 
         this.setState({
-          error: false,
+          error: null,
           issuesLoading: false,
           queryCount:
-            typeof queryCount !== 'undefined' ? parseInt(queryCount, 10) || 0 : 0,
+            typeof queryCount !== 'undefined' && queryCount
+              ? parseInt(queryCount, 10) || 0
+              : 0,
           queryMaxCount:
-            typeof queryMaxCount !== 'undefined' ? parseInt(queryMaxCount, 10) || 0 : 0,
-          pageLinks,
+            typeof queryMaxCount !== 'undefined' && queryMaxCount
+              ? parseInt(queryMaxCount, 10) || 0
+              : 0,
+          pageLinks: pageLinks !== null ? pageLinks : '',
         });
       },
       error: err => {
@@ -372,14 +409,14 @@ const IssueListOverview = createReactClass({
         });
       },
       complete: () => {
-        this.lastRequest = null;
+        this._lastRequest = null;
 
         this.resumePolling();
       },
     });
-  },
+  };
 
-  resumePolling() {
+  resumePolling = () => {
     if (!this.state.pageLinks) {
       return;
     }
@@ -390,41 +427,43 @@ const IssueListOverview = createReactClass({
       this._poller.setEndpoint(links.previous.href);
       this._poller.enable();
     }
-  },
+  };
 
-  getGroupListEndpoint() {
+  getGroupListEndpoint(): string {
     const params = this.props.params;
 
     return `/organizations/${params.orgId}/issues/`;
-  },
+  }
 
-  onRealtimeChange(realtime) {
+  onRealtimeChange = (realtime: boolean) => {
     Cookies.set('realtimeActive', realtime.toString());
     this.setState({
       realtimeActive: realtime,
     });
-  },
+  };
 
-  onSelectStatsPeriod(period) {
+  onSelectStatsPeriod = (period: string) => {
     if (period !== this.getGroupStatsPeriod()) {
       this.transitionTo({groupStatsPeriod: period});
     }
-  },
+  };
 
-  onRealtimePoll(data, _links) {
+  onRealtimePoll = (data: any, _links: any) => {
     // Note: We do not update state with cursors from polling,
     // `CursorPoller` updates itself with new cursors
     this._streamManager.unshift(data);
-  },
+  };
+
+  listener = GroupStore.listen(() => this.onGroupChange(), undefined);
 
   onGroupChange() {
     const groupIds = this._streamManager.getAllItems().map(item => item.id);
     if (!isEqual(groupIds, this.state.groupIds)) {
       this.setState({groupIds});
     }
-  },
+  }
 
-  onIssueListSidebarSearch(query) {
+  onIssueListSidebarSearch = (query: string) => {
     analytics('search.searched', {
       org_id: this.props.organization.id,
       query,
@@ -433,9 +472,9 @@ const IssueListOverview = createReactClass({
     });
 
     this.onSearch(query);
-  },
+  };
 
-  onSearch(query) {
+  onSearch = (query: string) => {
     if (query === this.state.query) {
       // if query is the same, just re-fetch data
       this.fetchData();
@@ -443,15 +482,17 @@ const IssueListOverview = createReactClass({
       // Clear the saved search as the user wants something else.
       this.transitionTo({query}, null);
     }
-  },
+  };
 
-  onSortChange(sort) {
+  onSortChange = (sort: string) => {
     this.transitionTo({sort});
-  },
+  };
 
-  onCursorChange(cursor, _path, query, pageDiff) {
+  onCursorChange = (cursor: string | undefined, _path, query, pageDiff: number) => {
     const queryPageInt = parseInt(query.page, 10);
-    let nextPage = isNaN(queryPageInt) ? pageDiff : queryPageInt + pageDiff;
+    let nextPage: number | undefined = isNaN(queryPageInt)
+      ? pageDiff
+      : queryPageInt + pageDiff;
 
     // unset cursor and page when we navigate back to the first page
     // also reset cursor if somehow the previous button is enabled on
@@ -462,9 +503,9 @@ const IssueListOverview = createReactClass({
     }
 
     this.transitionTo({cursor, page: nextPage});
-  },
+  };
 
-  onSidebarToggle() {
+  onSidebarToggle = () => {
     const {organization} = this.props;
     this.setState({
       isSidebarVisible: !this.state.isSidebarVisible,
@@ -472,27 +513,30 @@ const IssueListOverview = createReactClass({
     analytics('issue.search_sidebar_clicked', {
       org_id: parseInt(organization.id, 10),
     });
-  },
+  };
 
   /**
    * Returns true if all results in the current query are visible/on this page
    */
-  allResultsVisible() {
+  allResultsVisible(): boolean {
     if (!this.state.pageLinks) {
       return false;
     }
 
     const links = parseLinkHeader(this.state.pageLinks);
     return links && !links.previous.results && !links.next.results;
-  },
+  }
 
-  transitionTo(newParams = {}, savedSearch = this.props.savedSearch) {
+  transitionTo = (
+    newParams: Partial<EndpointParams> = {},
+    savedSearch: (SavedSearch & {projectId?: number}) | null = this.props.savedSearch
+  ) => {
     const query = {
       ...this.getEndpointParams(),
       ...newParams,
     };
     const {organization} = this.props;
-    let path;
+    let path: string;
 
     if (savedSearch && savedSearch.id) {
       path = `/organizations/${organization.slug}/issues/searches/${savedSearch.id}/`;
@@ -510,16 +554,19 @@ const IssueListOverview = createReactClass({
       path = `/organizations/${organization.slug}/issues/`;
     }
 
-    if (path !== this.props.location.path && !isEqual(query, this.props.location.query)) {
+    if (
+      path !== this.props.location.pathname ||
+      !isEqual(query, this.props.location.query)
+    ) {
       browserHistory.push({
         pathname: path,
         query,
       });
       this.setState({issuesLoading: true});
     }
-  },
+  };
 
-  renderGroupNodes(ids, groupStatsPeriod) {
+  renderGroupNodes = (ids: string[], groupStatsPeriod: string) => {
     const topIssue = ids[0];
     const {memberList} = this.state;
 
@@ -527,7 +574,7 @@ const IssueListOverview = createReactClass({
     const groupNodes = ids.map(id => {
       const hasGuideAnchor = id === topIssue;
       const group = GroupStore.get(id);
-      let members = null;
+      let members: Member['user'][] | null = null;
       if (group && group.project) {
         members = memberList[group.project.slug] || null;
       }
@@ -546,14 +593,14 @@ const IssueListOverview = createReactClass({
       );
     });
     return <PanelBody>{groupNodes}</PanelBody>;
-  },
+  };
 
-  renderLoading() {
+  renderLoading(): React.ReactNode {
     return <LoadingIndicator />;
-  },
+  }
 
-  renderStreamBody() {
-    let body;
+  renderStreamBody(): React.ReactNode {
+    let body: React.ReactNode;
     if (this.state.issuesLoading) {
       body = this.renderLoading();
     } else if (this.state.error) {
@@ -563,7 +610,7 @@ const IssueListOverview = createReactClass({
     } else {
       body = (
         <NoGroupsHandler
-          api={this.api}
+          api={this.props.api}
           organization={this.props.organization}
           query={this.getQuery()}
           selectedProjectIds={this.props.selection.projects}
@@ -572,27 +619,22 @@ const IssueListOverview = createReactClass({
       );
     }
     return body;
-  },
+  }
 
-  fetchSavedSearches() {
+  fetchSavedSearches = () => {
     const {organization} = this.props;
 
-    fetchSavedSearches(this.api, organization.slug);
-  },
+    fetchSavedSearches(this.props.api, organization.slug);
+  };
 
-  onSavedSearchCreate(newSavedSearch) {
-    // Navigate to new saved search
-    this.transitionTo(null, newSavedSearch);
-  },
+  onSavedSearchSelect = (savedSearch: SavedSearch) => {
+    this.setState({issuesLoading: true}, () => this.transitionTo(undefined, savedSearch));
+  };
 
-  onSavedSearchSelect(savedSearch) {
-    this.setState({issuesLoading: true}, () => this.transitionTo(null, savedSearch));
-  },
-
-  onSavedSearchDelete(search) {
+  onSavedSearchDelete = (search: SavedSearch) => {
     const {orgId} = this.props.params;
 
-    deleteSavedSearch(this.api, orgId, search).then(() => {
+    deleteSavedSearch(this.props.api, orgId, search).then(() => {
       this.setState(
         {
           issuesLoading: true,
@@ -600,15 +642,22 @@ const IssueListOverview = createReactClass({
         () => this.transitionTo({}, null)
       );
     });
-  },
+  };
 
-  tagValueLoader(key, search) {
+  tagValueLoader = (key: string, search: string) => {
     const {orgId} = this.props.params;
-    const projectIds = this.getGlobalSearchProjectIds();
+    const projectIds = this.getGlobalSearchProjectIds().map(id => id.toString());
     const endpointParams = this.getEndpointParams();
 
-    return fetchTagValues(this.api, orgId, key, search, projectIds, endpointParams);
-  },
+    return fetchTagValues(
+      this.props.api,
+      orgId,
+      key,
+      search,
+      projectIds,
+      endpointParams as any
+    );
+  };
 
   render() {
     if (this.props.savedSearchLoading) {
@@ -619,7 +668,7 @@ const IssueListOverview = createReactClass({
       classes.push('show-sidebar');
     }
 
-    const {params, organization, savedSearch, savedSearches, tags} = this.props;
+    const {organization, savedSearch, savedSearches, tags} = this.props;
     const query = this.getQuery();
 
     return (
@@ -627,7 +676,6 @@ const IssueListOverview = createReactClass({
         <div className="stream-content">
           <IssueListFilters
             organization={organization}
-            searchId={params.searchId}
             query={query}
             savedSearch={savedSearch}
             sort={this.getSort()}
@@ -635,7 +683,6 @@ const IssueListOverview = createReactClass({
             queryMaxCount={this.state.queryMaxCount}
             onSortChange={this.onSortChange}
             onSearch={this.onSearch}
-            onSavedSearchCreate={this.onSavedSearchCreate}
             onSavedSearchSelect={this.onSavedSearchSelect}
             onSavedSearchDelete={this.onSavedSearchDelete}
             onSidebarToggle={this.onSidebarToggle}
@@ -662,7 +709,7 @@ const IssueListOverview = createReactClass({
             <PanelBody>
               <ProcessingIssueList
                 organization={this.props.organization}
-                projectIds={this.props.selection.projects}
+                projectIds={this.props.selection?.projects?.map(p => p.toString())}
                 showProject
               />
               {this.renderStreamBody()}
@@ -680,10 +727,12 @@ const IssueListOverview = createReactClass({
         />
       </div>
     );
-  },
-});
+  }
+}
 
-export default withGlobalSelection(
-  withSavedSearches(withOrganization(withIssueTags(withProfiler(IssueListOverview))))
+export default withApi(
+  withGlobalSelection(
+    withSavedSearches(withOrganization(withIssueTags(withProfiler(IssueListOverview))))
+  )
 );
 export {IssueListOverview};
