@@ -1,7 +1,3 @@
 import {TagValue} from 'app/types';
 
-export type TagValueLoader = (
-  key: string,
-  search: string,
-  projectIds?: string[]
-) => Promise<TagValue[]>;
+export type TagValueLoader = (key: string, search: string) => Promise<TagValue[]>;
